Fix extremely embarrassing bug thanks to @doomrobo
Bug causes sfs to upload a the filename as the image data. The problem
is that we're supplying filename as an argument to the method that does
the uploading -- but the argument should be the image data.

This commit will fix this issue by changing having the upload function
do the data reading.